backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
bcrypt
python-multipart
pydantic-settings
python-dotenv
cloudinary
